This is PBL 3
Hi this is the Problem Based Learning task of my third semester. This problem is taken from Kaggle website. 
In this problem data is saved so to get the information of the earth quack of different countries more over you can see
the problem in the file PBL(21SW).pdf
There may be some logical errors in the code I will thankful to you if you let me know by creating issue and making pull request